requirements.txt
pyyaml
requests
opencv-python
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.10.0+cu113
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torchvision==0.11.1+cu113
tqdm
scikit-learn
umap-learn
matplotlib